+++ app/src/main/java/com/egci428/shopmai/Model/Order.kt
@@ -1,4 +1,4 @@
 package com.egci428.shopmai.Model
 
-data class Order (val id: Int, val title: String, val img1: String, val price: Int)
+data class Order (val id: Int, val title: String, val img1: String, val price: Float)
